+++ CHANGELOG.md
@@ -1,2 +1,1 @@
-- Feature: Added step to clean up job quests/other priority quests before continuing with msq -alydev
-- Change: crystal icon now always visible, hovering shows available job quests/other priority quests. grey when empty, amber when not -alydev
+- Feature: class quests and quests that can be done by only one class now check that class's level (instead of the player's current job's level) when testing if locked -alydev